combine tests, remove `should` from deps
